finished

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: aadamik <aadamik@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 20:02:58 by aadamik           #+#    #+#             */
-/*   Updated: 2023/11/15 17:07:10 by aadamik          ###   ########.fr       */
+/*   Updated: 2023/11/17 00:00:15 by aadamik          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
