+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: aadamik <aadamik@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/16 19:59:30 by aadamik           #+#    #+#             */
-/*   Updated: 2023/11/16 22:04:46 by aadamik          ###   ########.fr       */
+/*   Updated: 2023/11/17 01:43:46 by aadamik          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
-	int	k;
+	size_t	i;
+	size_t	j;
+	size_t	k;
 
 	i = -1;
 	if (needle[0] == '\0')
